+++ src/main/java/com/awesometeam/Invoicify/invoice/model/Invoice.java
@@ -1,0 +1,4 @@
+package com.awesometeam.Invoicify.invoice.model;
+
+public class Invoice {
+}
