+++ src/main/java/com/awesometeam/Invoicify/invoice/repository/InvoiceRepository.java
@@ -1,0 +1,2 @@
+package com.awesometeam.Invoicify.invoice.repository;public interface InvoiceRepository {
+}
