+++ src/main/java/com/awesometeam/Invoicify/invoice/service/InvoiceService.java
@@ -1,0 +1,2 @@
+package com.awesometeam.Invoicify.invoice.service;public class InvoiceService {
+}
