+++ src/test/java/com/awesometeam/Invoicify/invoice/integration/InvoiceIntegrationTests.java
@@ -1,0 +1,2 @@
+package com.awesometeam.Invoicify.invoice.integration;public class InvoiceIntegrationTests {
+}
