+++ src/test/java/com/awesometeam/Invoicify/invoice/unit/InvoiceApiTests.java
@@ -1,0 +1,2 @@
+package com.awesometeam.Invoicify.invoice.unit;public class InvoiceApiTests {
+}
