+++ src/test/java/com/awesometeam/Invoicify/invoice/unit/InvoiceServiceTests.java
@@ -1,0 +1,4 @@
+package com.awesometeam.Invoicify.invoice.unit;
+
+public class InvoiceServiceTests {
+}
